+++ src/gpre/ftn.cpp
@@ -24,7 +24,7 @@
 //
 //____________________________________________________________
 //
-//	$Id: ftn.cpp,v 1.20.2.1 2005-10-20 13:01:30 kkuznetsov Exp $
+//	$Id: ftn.cpp,v 1.20.2.2 2005-10-24 16:51:47 awharrison Exp $
 //
 // 2002.10.28 Sean Leyne - Completed removal of obsolete "DGUX" port
 // 2002.10.28 Sean Leyne - Completed removal of obsolete "SGI" port
@@ -2466,7 +2466,7 @@ static void gen_event_wait( ACT action)
 
 	args.pat_database = database;
 	args.pat_vector1 = status_vector(action);
-	args.pat_value1 = (int) ident;
+	args.pat_value1 = ident;
 	args.pat_string1 = ISC_EVENT_WAIT;
 	args.pat_string2 = ISC_EVENT_COUNTS;
